refactor: rename webserver module
