+++ app.yaml
@@ -8,4 +8,4 @@ env_variables:
 
 service_account: not-that-movie-worker@webhost-common.iam.gserviceaccount.com
 
-entrypoint: gunicorn -b :$PORT --timeout 300 app.views:app
+entrypoint: gunicorn -b :$PORT --timeout 300 webserver.views:app
