+++ pyproject.toml
@@ -1,7 +1,7 @@
 [project]
 name = "not_that_movie"
 version = "1.0.0"
-description = "Movie alternative recommender app"
+description = "Webapp for generating badly translated movie descriptions."
 requires-python = ">=3.11"
 dependencies = [
     "beautifulsoup4",
