+++ tests/test_utils.py
@@ -1,6 +1,6 @@
 import pytest
 
-from app import (
+from webserver import (
 	utils
 )
