+++ webserver/data/weight_config.json
