+++ webserver/static/favicon.ico
